Core Utilities
I've been writing this replacement for 'coreutils' programs for a few
months, without the assistance of (but always heading towards using) version
control. So a few of the utilities have been written.

Note: am isn't part of the ordinary coreutils, but is still core to
these utilities.

Each source file 'includes' a modified zlib/libpng license.

Signed-off-by: Daniel Cordero <theappleman@gmail.com>