Be compatible with other compilers
This code was originally written for use with tinycc (tcc), but it is
still light enough to be compatible with other compilers, with the
exception of header files.

tcc, gcc and icc should be able to use these files with no problem.

The Plan 9 compliers should also work with little change to the code, so
their header files are also included. However, Plan 9 needs an explicit
linker/loader stage, but automatic instructions have not been included.

note: future utilities may not be compatible with all compilers, so I may
not keep this feature in the future.

Signed-off-by: Daniel Cordero <theappleman@gmail.com>