Make the license prettier
I completely expect people to edit the license to add their name if
they were to edit the source code (in fact, adding your name also edits
the source). One license series (not in effect) mandates thatyou do
such a thing.

The output of `strings <exe>` did not look particularly nice, so put
some newlines in.

With regard to version control systems `annotate`, the last person will
touch the line above, so try to not move the semi-colon or touch anyone
else's line. Put a newline to separate it out when viewed in compiled
(not-quite-binary form).

Signed-off-by: Daniel Cordero <theappleman@gmail.com>