mkfile: use meta-rules
mk(1) supports meta-rules; it'll cut down the number of added lines when
adding a new, solo C-file.

make(1) also supports meta-rules, but it is not included in the makefile
because am(1) does not. make(1) has implicit meta-rules, including one
for C source files, so it makes little difference.

NOTE: a major difference is that the meta-rules do not respect changes
in license.h (it is not on the dependancy list). If you have touched it,
and want to recompile everything, use `mk clean all`.

Signed-off-by: Daniel Cordero <theappleman@gmail.com>