echo: check that argv is sane
The test in the while() loop incrementing argv will see the variable as
an integer (the location of the first string). It's just a number, but
doesn't reliably say the string is valid.

Instead, check that the beginning of the string is NOT null (it doesn't
matter too much if it doesn't end...)

Signed-off-by: Daniel Cordero <theappleman@gmail.com>